Change (break) API, to receive binary data instead of file handles.
The original reason for receiving a handle was that the binary format requires a lot of seeking,
and using a file handle could be more RAM-friendly.
However, I think it’s realistically unlikely that a plist will be large enough for this to be a concern,
and other parsing libraries (e.g. Poison) provide only the single parsing entry point,
so I think this is the best choice.